Add is_local on Series
