+++ py-polars/docs/source/reference/series/categories.rst
@@ -10,6 +10,7 @@ The following methods are available under the `Series.cat` attribute.
    :template: autosummary/accessor_method.rst
 
     Series.cat.get_categories
+    Series.cat.is_local
     Series.cat.set_ordering
     Series.cat.to_local
     Series.cat.uses_lexical_ordering
